pyproject.toml
[tool.poetry]
name = "em-observability-backend"
version = "0.1.0"
description = ""
authors = ["pardhu21 <pardhuganesh21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.24.0"
fastapi = "^0.104.1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
pymongo = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
